feat: add "cluster attach" command
